Refactor the MQTT driver
1. Split source code files
2. Add copyright information
3. Split parser into different files
