+++ plugins/mqtt/CMakeLists.txt
@@ -4,7 +4,7 @@ find_package(eclipse-paho-mqtt-c 1.3.9 REQUIRED)
 
 file(GLOB PARSER_SRC "${CMAKE_SOURCE_DIR}/src/parser/*.c")
 set(LIBRARY_OUTPUT_PATH "${CMAKE_BINARY_DIR}")
-set(src ${PARSER_SRC} plugin_mqtt.c paho_client.c)
+set(src ${PARSER_SRC} plugin_mqtt.c paho_client.c message_handle.c)
 add_library(plugin-mqtt SHARED ${src})
 
 target_include_directories(plugin-mqtt PRIVATE
