+++ plugins/mqtt/paho_client.h
@@ -1,3 +1,22 @@
+/**
+ * NEURON IIoT System for Industry 4.0
+ * Copyright (C) 2020-2021 EMQ Technologies Co., Ltd All rights reserved.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ **/
+
 #ifndef NEURON_PLUGIN_PAHO_CLIENT
 #define NEURON_PLUGIN_PAHO_CLIENT
 
@@ -54,6 +73,8 @@ client_error paho_client_unsubscribe(paho_client_t *client, const char *topic);
 
 client_error paho_client_publish(paho_client_t *client, const char *topic,
                                  int qos, unsigned char *payload, size_t len);
+client_error paho_client_publish2(paho_client_t *client, const char *topic,
+                                  int qos, unsigned char *payload, size_t len);
 
 client_error paho_client_close(paho_client_t *client);
