+++ src/parser/neu_json_tag_group.c
@@ -18,10 +18,11 @@
  **/
 
 #include <stdlib.h>
+#include <string.h>
 
 #include "utils/json.h"
 
-#include "neu_json_read.h"
+#include "neu_json_tag_group.h"
 
 #define GROUP "group"
 #define NAMES "names"
@@ -31,16 +32,19 @@
 #define VALUE "value"
 #define TAGS "tags"
 
-int neu_parse_decode_read_req(char *buf, struct neu_parse_read_req **result)
+int neu_paser_decode_read_tag_group_list_req(
+    char *buf, struct neu_paser_read_tag_group_list_req **result)
 {
-    struct neu_parse_read_req *req =
-        calloc(1, sizeof(struct neu_parse_read_req));
+    struct neu_paser_read_tag_group_list_req *req =
+        malloc(sizeof(struct neu_paser_read_tag_group_list_req));
+    memset(req, 0, sizeof(struct neu_paser_read_tag_group_list_req));
+
     neu_json_elem_t elem[] = { {
                                    .name = NEU_PARSE_UUID,
                                    .t    = NEU_JSON_STR,
                                },
                                {
-                                   .name = GROUP,
+                                   .name = "config",
                                    .t    = NEU_JSON_STR,
                                }
 
@@ -51,56 +55,31 @@ int neu_parse_decode_read_req(char *buf, struct neu_parse_read_req **result)
         free(req);
         return -1;
     }
-    req->function = NEU_PARSE_OP_READ;
+    req->function = NEU_PARSE_OP_READ_TAG_GROUP_LIST;
     req->uuid     = elem[0].v.val_str;
-    req->group    = elem[1].v.val_str;
-
-    req->n_name = neu_json_decode_array_size(buf, NAMES);
-    req->names  = calloc(req->n_name, sizeof(struct neu_parse_read_req_name));
-    for (int i = 0; i < req->n_name; i++) {
-        neu_json_elem_t name = {
-            .name = NULL,
-            .t    = NEU_JSON_STR,
-        };
-        neu_json_decode_array(buf, NAMES, i, 1, &name);
-        req->names[i].name = name.v.val_str;
-    }
+    req->config   = elem[1].v.val_str;
 
     *result = req;
-
     return 0;
 }
-int neu_parse_encode_read_res(struct neu_parse_read_res *res, char **buf)
+
+int neu_parse_encode_read_tag_group_list_res(
+    struct neu_paser_read_tag_group_list_res *res, char **buf)
 {
     void *array = NULL;
-    for (int i = 0; i < res->n_tag; i++) {
-        neu_json_elem_t tag[] = { {
-                                      .name      = NAME,
-                                      .t         = NEU_JSON_STR,
-                                      .v.val_str = res->tags[i].name,
-                                  },
-                                  {
-                                      .name      = TYPE,
-                                      .t         = NEU_JSON_INT,
-                                      .v.val_int = res->tags[i].type,
-
-                                  },
-                                  {
-                                      .name      = TIMESTAMP,
-                                      .t         = NEU_JSON_INT,
-                                      .v.val_int = res->tags[i].timestamp,
-                                  },
-                                  {
-                                      .name         = VALUE,
-                                      .t            = NEU_JSON_DOUBLE,
-                                      .v.val_double = res->tags[i].value,
-                                  } };
-        array                 = neu_json_encode_array(array, tag, 4);
+    for (int i = 0; i < res->n_group; i++) {
+        neu_json_elem_t group[] = { {
+            .name      = NAME,
+            .t         = NEU_JSON_STR,
+            .v.val_str = res->names[i].name,
+        } };
+        array                   = neu_json_encode_array(array, group, 1);
     }
     neu_json_elem_t elems[] = { {
-                                    .name      = NEU_PARSE_FUNCTION,
-                                    .t         = NEU_JSON_INT,
-                                    .v.val_int = NEU_PARSE_OP_READ,
+                                    .name = NEU_PARSE_FUNCTION,
+                                    .t    = NEU_JSON_INT,
+                                    .v.val_int =
+                                        NEU_PARSE_OP_READ_TAG_GROUP_LIST,
                                 },
                                 {
                                     .name      = NEU_PARSE_UUID,
@@ -113,10 +92,9 @@ int neu_parse_encode_read_res(struct neu_parse_read_res *res, char **buf)
                                     .v.val_int = res->error,
                                 },
                                 {
-                                    .name     = TAGS,
+                                    .name     = "groups",
                                     .t        = NEU_JSON_OBJECT,
                                     .v.object = array,
                                 } };
-
     return neu_json_encode(elems, 4, buf);
 }
